philos

+++ philo/src/millis.c
@@ -6,18 +6,18 @@
 /*   By: brunofer <brunofer@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/26 16:33:04 by brunofer          #+#    #+#             */
-/*   Updated: 2025/10/26 16:38:59 by brunofer         ###   ########.fr       */
+/*   Updated: 2025/10/29 18:50:29 by brunofer         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "millis.h"
 
-unsigned long	millis(void)
+long long	millis(void)
 {
 	struct timeval	timeofday;
-	unsigned long	sec_to_millis;
-	unsigned long	micro_to_millis;
-	unsigned long	millis;
+	long long		sec_to_millis;
+	long long		micro_to_millis;
+	long long		millis;
 
 	gettimeofday(&timeofday, 0);
 	sec_to_millis = timeofday.tv_sec * 1000;
